ingestion/event_processor/requirements.txt
azure-functions
azure-cosmos>=4.5.0
httpx>=0.27.0
openai>=1.30.0
azure-iot-device>=2.14.0
